slam,good cloud, no merge cloud

+++ src/slamBase.cpp
@@ -19,7 +19,7 @@ PointCloud::Ptr image2PointCloud( cv::Mat& rgb, cv::Mat& depth, CAMERA_INTRINSIC
         for (int n=0; n < depth.cols; n++)
         {
             // 获取深度图中(m,n)处的值
-            ushort d = depth.ptr<ushort>(m)[n];
+            uchar d = depth.ptr<uchar>(m)[n];
             // d 可能没有值，若如此，跳过此点
             if (d == 0)
                 continue;
@@ -129,7 +129,7 @@ static ParameterReader pd;
         // query 是第一个, train 是第二个
         cv::Point2f p = frame1.kp[goodMatches[i].queryIdx].pt;
         // 获取d是要小心！x是向右的，y是向下的，所以y才是行，x是列！
-        ushort d = frame1.depth.ptr<ushort>( int(p.y) )[ int(p.x) ];
+        uchar d = frame1.depth.ptr<uchar>( int(p.y) )[ int(p.x) ];
         if (d == 0)
             continue;
         pts_img.push_back( cv::Point2f( frame2.kp[goodMatches[i].trainIdx].pt ) );
